Making the Sample Editor pane work

+++ Thumper - Leaf Editor/ThumperEditor.cs
@@ -47,6 +47,7 @@ namespace Thumper___Leaf_Editor
                     LvlEditorVisible(false); SaveLvl(true);
                     GateEditorVisible(false); SaveGate(true);
                     MasterEditorVisible(false); SaveMaster(true);
+                    SampleEditorVisible(false); SaveSample(true);
 
                     //set Working Folder panel data
                     btnWorkRefresh_Click(null, null);
@@ -180,6 +181,7 @@ namespace Thumper___Leaf_Editor
             InitializeLvlStuff();
             InitializeMasterStuff();
             InitializeGateStuff();
+            InitializeSampleStuff();
             //set panels to be resizeable
             ControlMoverOrResizer.Init(panelLeaf);
             ControlMoverOrResizer.Init(panelLevel);
@@ -199,6 +201,7 @@ namespace Thumper___Leaf_Editor
                 File.WriteAllText(@"templates\leaf_singletrack.txt", Properties.Resources.leaf_singletrack);
                 File.WriteAllText(@"templates\leaf_multitrack.txt", Properties.Resources.leaf_multitrack);
             }
+            //call method that imports objects from track_objects.txt (for Leaf editing)
             ImportObjects();
 
             //set timesig datasource
@@ -325,91 +328,91 @@ namespace Thumper___Leaf_Editor
             workingfolder = input.txtCustomPath.Text;
             ///create samp_ files if any boxes are checked
             //level 1
-            if (input.chkLevel1.Checked) {
+            if (input.chkLevel1.Checked && !File.Exists($@"{workingfolder}\samp_level1_320bpm.txt")) {
                 File.WriteAllText($@"{workingfolder}\samp_level1_320bpm.txt", Properties.Resources.samp_level1_320bpm);
             }
             else if (File.Exists($@"{workingfolder}\samp_level1_320bpm.txt")) {
                 File.Delete($@"{workingfolder}\samp_level1_320bpm.txt");
             }
             //level 2
-            if (input.chkLevel2.Checked) {
+            if (input.chkLevel2.Checked && !File.Exists($@"{workingfolder}\samp_level2_340bpm.txt")) {
                 File.WriteAllText($@"{workingfolder}\samp_level2_340bpm.txt", Properties.Resources.samp_level2_340bpm);
             }
             else if (File.Exists($@"{workingfolder}\samp_level2_340bpm.txt")) {
                 File.Delete($@"{workingfolder}\samp_level2_340bpm.txt");
             }
             //level 3
-            if (input.chkLevel3.Checked) {
+            if (input.chkLevel3.Checked && !File.Exists($@"{workingfolder}\samp_level3_360bpm.txt")) {
                 File.WriteAllText($@"{workingfolder}\samp_level3_360bpm.txt", Properties.Resources.samp_level3_360bpm);
             }
             else if (File.Exists($@"{workingfolder}\samp_level3_360bpm.txt")) {
                 File.Delete($@"{workingfolder}\samp_level3_360bpm.txt");
             }
             //level 4
-            if (input.chkLevel4.Checked) {
+            if (input.chkLevel4.Checked && !File.Exists($@"{workingfolder}\samp_level4_380bpm.txt")) {
                 File.WriteAllText($@"{workingfolder}\samp_level4_380bpm.txt", Properties.Resources.samp_level4_380bpm);
             }
             else if (File.Exists($@"{workingfolder}\samp_level4_380bpm.txt")) {
                 File.Delete($@"{workingfolder}\samp_level4_380bpm.txt");
             }
             //level 5
-            if (input.chkLevel5.Checked) {
+            if (input.chkLevel5.Checked && !File.Exists($@"{workingfolder}\samp_level5_400bpm.txt")) {
                 File.WriteAllText($@"{workingfolder}\samp_level5_400bpm.txt", Properties.Resources.samp_level5_400bpm);
             }
             else if (File.Exists($@"{workingfolder}\samp_level5_400bpm.txt")) {
                 File.Delete($@"{workingfolder}\samp_level5_400bpm.txt");
             }
             //level 6
-            if (input.chkLevel6.Checked) {
+            if (input.chkLevel6.Checked && !File.Exists($@"{workingfolder}\samp_level6_420bpm.txt")) {
                 File.WriteAllText($@"{workingfolder}\samp_level6_420bpm.txt", Properties.Resources.samp_level6_420bpm);
             }
             else if (File.Exists($@"{workingfolder}\samp_level6_420bpm.txt")) {
                 File.Delete($@"{workingfolder}\samp_level6_420bpm.txt");
             }
             //level 7
-            if (input.chkLevel7.Checked) {
+            if (input.chkLevel7.Checked && !File.Exists($@"{workingfolder}\samp_level7_440bpm.txt")) {
                 File.WriteAllText($@"{workingfolder}\samp_level7_440bpm.txt", Properties.Resources.samp_level7_440bpm);
             }
             else if (File.Exists($@"{workingfolder}\samp_level7_440bpm.txt")) {
                 File.Delete($@"{workingfolder}\samp_level7_440bpm.txt");
             }
             //level 8
-            if (input.chkLevel8.Checked) {
+            if (input.chkLevel8.Checked && !File.Exists($@"{workingfolder}\samp_level8_460bpm.txt")) {
                 File.WriteAllText($@"{workingfolder}\samp_level8_460bpm.txt", Properties.Resources.samp_level8_460bpm);
             }
             else if (File.Exists($@"{workingfolder}\samp_level8_460bpm.txt")) {
                 File.Delete($@"{workingfolder}\samp_level8_460bpm.txt");
             }
             //level 9
-            if (input.chkLevel9.Checked) {
+            if (input.chkLevel9.Checked && !File.Exists($@"{workingfolder}\samp_level9_480bpm.txt")) {
                 File.WriteAllText($@"{workingfolder}\samp_level9_480bpm.txt", Properties.Resources.samp_level9_480bpm);
             }
             else if (File.Exists($@"{workingfolder}\samp_level9_480bpm.txt")) {
                 File.Delete($@"{workingfolder}\samp_level9_480bpm.txt");
             }
             //Dissonance
-            if (input.chkDissonance.Checked) {
+            if (input.chkDissonance.Checked && !File.Exists($@"{workingfolder}\samp_dissonant.txt")) {
                 File.WriteAllText($@"{workingfolder}\samp_dissonant.txt", Properties.Resources.samp_dissonant);
             }
             else if (File.Exists($@"{workingfolder}\samp_dissonant.txt")) {
                 File.Delete($@"{workingfolder}\samp_dissonant.txt");
             }
             //Global Drones
-            if (input.chkGlobal.Checked) {
+            if (input.chkGlobal.Checked && !File.Exists($@"{workingfolder}\samp_globaldrones.txt")) {
                 File.WriteAllText($@"{workingfolder}\samp_globaldrones.txt", Properties.Resources.samp_globaldrones);
             }
             else if (File.Exists($@"{workingfolder}\samp_globaldrones.txt")) {
                 File.Delete($@"{workingfolder}\samp_globaldrones.txt");
             }
             //Rests
-            if (input.chkRests.Checked) {
+            if (input.chkRests.Checked && !File.Exists($@"{workingfolder}\samp_rests.txt")) {
                 File.WriteAllText($@"{workingfolder}\samp_rests.txt", Properties.Resources.samp_rests);
             }
             else if (File.Exists($@"{workingfolder}\samp_rests.txt")) {
                 File.Delete($@"{workingfolder}\samp_rests.txt");
             }
             //Misc
-            if (input.chkMisc.Checked) {
+            if (input.chkMisc.Checked && !File.Exists($@"{workingfolder}\samp_misc.txt")) {
                 File.WriteAllText($@"{workingfolder}\samp_misc.txt", Properties.Resources.samp_misc);
             }
             else if (File.Exists($@"{workingfolder}\samp_misc.txt")) {
@@ -508,16 +511,6 @@ namespace Thumper___Leaf_Editor
                 ColorFormElements();
             }
             custom.Dispose();
-        }
-
-        private void button1_Click(object sender, EventArgs e)
-        {
-            /*
-			byte[] bytes = BitConverter.GetBytes(Hash32("Asamples/levels/custom/rose.wav"));
-			Array.Reverse(bytes);
-			foreach (byte b in bytes)
-				lbltest.Text += b.ToString("X").PadLeft(2, '0').ToLower();
-			*/
         }
     }
 }
